+++ hslua-packaging/src/HsLua/Packaging/Documentation.hs
@@ -13,6 +13,7 @@ module HsLua.Packaging.Documentation
   , registerDocumentation
   , pushModuleDoc
   , pushFunctionDoc
+  , pushFieldDoc
   , docsField
   ) where