+++ hslua-packaging/test/HsLua/Packaging/ModuleTests.hs
@@ -12,8 +12,11 @@ Tests creating and loading of modules with Haskell.
 -}
 module HsLua.Packaging.ModuleTests (tests) where
 
+import HsLua.Core
 import HsLua.Marshalling
-  (Result (Success), peekIntegral, peekString, pushIntegral, pushText, runPeek)
+  ( forcePeek, peekFieldRaw, peekIntegral, peekString
+  , pushIntegral, pushText)
+import HsLua.Packaging.Documentation
 import HsLua.Packaging.Function
 import HsLua.Packaging.Module
 import Test.Tasty.HsLua ((=:), shouldBeResultOf)
@@ -40,22 +43,62 @@ tests = testGroup "Module"
         return (new - old)
 
     , "call module function" =:
-      Success 24 `shouldBeResultOf` do
+      24 `shouldBeResultOf` do
         Lua.openlibs
         registerModule mymath
         _ <- Lua.dostring $ mconcat
              [ "local mymath = require 'mymath'\n"
              , "return mymath.factorial(4)"
              ]
-        runPeek $ peekIntegral @Integer Lua.top
+        forcePeek $ peekIntegral @Prelude.Integer Lua.top
 
     , "call module as function" =:
-      Success "call me maybe" `shouldBeResultOf` do
+      "call me maybe" `shouldBeResultOf` do
         Lua.openlibs
         registerModule mymath
         _ <- Lua.dostring "return (require 'mymath')()"
-        runPeek $ peekString Lua.top
+        forcePeek $ peekString Lua.top
 
+    , "access name in docs" =:
+      "mymath" `shouldBeResultOf` do
+        Lua.openlibs
+        registerModule mymath
+        TypeTable <- getdocumentation top
+        forcePeek $ peekFieldRaw peekString "name" Lua.top
+
+    , "first function name in docs" =:
+      "factorial" `shouldBeResultOf` do
+        Lua.openlibs
+        registerModule mymath
+        TypeTable <- getdocumentation top
+        TypeTable <- getfield top "functions"
+        TypeTable <- rawgeti top 1
+        forcePeek $ peekFieldRaw peekString "name" Lua.top
+
+    , "function doc is shared" =:
+      True `shouldBeResultOf` do
+        Lua.openlibs
+        registerModule mymath
+        pushvalue top
+        setglobal "mymath"
+        -- get doc table via module docs
+        TypeTable <- getdocumentation top
+        TypeTable <- getfield top "functions"
+        TypeTable <- rawgeti top 1
+        -- get doc table via function
+        OK <- dostring "return mymath.factorial"
+        TypeTable <- getdocumentation top
+        -- must be the same
+        rawequal (nth 1) (nth 3)
+
+    , "first field name in docs" =:
+      "unit" `shouldBeResultOf` do
+        Lua.openlibs
+        registerModule mymath
+        TypeTable <- getdocumentation top
+        TypeTable <- getfield top "fields"
+        TypeTable <- rawgeti top 1
+        forcePeek $ peekFieldRaw peekString "name" Lua.top
     ]
   ]
 
@@ -63,7 +106,9 @@ mymath :: Module Lua.Exception
 mymath = Module
   { moduleName = "mymath"
   , moduleDescription = "A math module."
-  , moduleFields = []
+  , moduleFields = [
+      Field "unit" "additive unit" (pushinteger 1)
+    ]
   , moduleFunctions = [factorial]
   , moduleOperations =
     [ (,) Call $ lambda
@@ -79,12 +124,11 @@ factorial =
   <#> factorialParam
   =#> factorialResult
 
-factorialParam :: Parameter Lua.Exception Integer
+factorialParam :: Parameter Lua.Exception Prelude.Integer
 factorialParam =
-  parameter (peekIntegral @Integer) "integer"
+  parameter peekIntegral "integer"
     "n"
     "number for which the factorial is computed"
 
-factorialResult :: FunctionResults Lua.Exception Integer
-factorialResult =
-  functionResult (pushIntegral @Integer) "integer" "factorial"
+factorialResult :: FunctionResults Lua.Exception Prelude.Integer
+factorialResult = functionResult pushIntegral "integer" "factorial"